chore: rename core test scripts, fix config
