+++ packages/core/test/UserState/userStateTransitionEvent.test.ts
@@ -10,16 +10,18 @@ import {
     computeStartTransitionProofHash,
     computeProcessAttestationsProofHash,
 } from '@unirep/contracts'
+import { formatProofForVerifierContract } from '@unirep/circuits'
 import {
-    attestingFee,
+    EPOCH_LENGTH,
+    MAX_REPUTATION_BUDGET,
+    NUM_EPOCH_KEY_NONCE_PER_EPOCH,
+} from '@unirep/config'
+
+import {
     computeInitUserStateRoot,
-    epochLength,
     genUnirepStateFromContract,
     genUserStateFromContract,
     ISettings,
-    maxAttesters,
-    maxReputationBudget,
-    numEpochKeyNoncePerEpoch,
     Reputation,
     UnirepState,
     UserState,
@@ -32,7 +34,6 @@ import {
     verifyProcessAttestationsProof,
     verifyStartTransitionProof,
 } from '../utils'
-import { formatProofForVerifierContract } from '@unirep/circuits'
 
 describe('User state transition events in Unirep User State', async function () {
     this.timeout(0)
@@ -45,13 +46,14 @@ describe('User state transition events in Unirep User State', async function () 
 
     let unirepContract: ethers.Contract
     let unirepContractCalledByAttester: ethers.Contract
-    let _treeDepths = getTreeDepthsForTesting('circuit')
+    let _treeDepths = getTreeDepthsForTesting()
 
     let accounts: ethers.Signer[]
     const attester = new Object()
     let attesterId
     const maxUsers = 10
     const userNum = Math.ceil(Math.random() * maxUsers)
+    const attestingFee = ethers.utils.parseEther('0.1')
     const transitionedUsers: number[] = []
     const fromProofIndex = 0
 
@@ -59,12 +61,8 @@ describe('User state transition events in Unirep User State', async function () 
         accounts = await hardhatEthers.getSigners()
 
         const _settings = {
-            maxUsers: maxUsers,
-            maxAttesters: maxAttesters,
-            numEpochKeyNoncePerEpoch: numEpochKeyNoncePerEpoch,
-            maxReputationBudget: maxReputationBudget,
-            epochLength: epochLength,
-            attestingFee: attestingFee,
+            maxUsers,
+            attestingFee,
         }
         unirepContract = await deployUnirep(
             <ethers.Wallet>accounts[0],
@@ -271,7 +269,9 @@ describe('User state transition events in Unirep User State', async function () 
             let epoch = await unirepContract.currentEpoch()
 
             // Fast-forward epochLength of seconds
-            await hardhatEthers.provider.send('evm_increaseTime', [epochLength])
+            await hardhatEthers.provider.send('evm_increaseTime', [
+                EPOCH_LENGTH,
+            ])
 
             // Begin epoch transition
             let tx = await unirepContractCalledByAttester.beginEpochTransition()
@@ -310,9 +310,9 @@ describe('User state transition events in Unirep User State', async function () 
             userStateTreeDepth: _treeDepths.userStateTreeDepth,
             epochTreeDepth: _treeDepths.epochTreeDepth,
             attestingFee: attestingFee,
-            epochLength: epochLength,
-            numEpochKeyNoncePerEpoch: numEpochKeyNoncePerEpoch,
-            maxReputationBudget: maxReputationBudget,
+            epochLength: EPOCH_LENGTH,
+            numEpochKeyNoncePerEpoch: NUM_EPOCH_KEY_NONCE_PER_EPOCH,
+            maxReputationBudget: MAX_REPUTATION_BUDGET,
         }
         it('Users should successfully perform user state transition', async () => {
             for (let i = 0; i < userIds.length; i++) {
@@ -606,11 +606,11 @@ describe('User state transition events in Unirep User State', async function () 
         it('Submit invalid user state transition proof should not affect Unirep State', async () => {
             const randomProof: BigInt[] = genRandomList(8)
             const randomNullifiers: BigInt[] = genRandomList(
-                numEpochKeyNoncePerEpoch
+                NUM_EPOCH_KEY_NONCE_PER_EPOCH
             )
             const randomBlindedStates: BigInt[] = genRandomList(2)
             const randomBlindedChains: BigInt[] = genRandomList(
-                numEpochKeyNoncePerEpoch
+                NUM_EPOCH_KEY_NONCE_PER_EPOCH
             )
 
             const randomUSTInput = {
@@ -889,7 +889,9 @@ describe('User state transition events in Unirep User State', async function () 
             let epoch = await unirepContract.currentEpoch()
 
             // Fast-forward epochLength of seconds
-            await hardhatEthers.provider.send('evm_increaseTime', [epochLength])
+            await hardhatEthers.provider.send('evm_increaseTime', [
+                EPOCH_LENGTH,
+            ])
             // Begin epoch transition
             let tx = await unirepContractCalledByAttester.beginEpochTransition()
             let receipt = await tx.wait()
